fix issue with chat messages being posted multiple times
